user role data group related changes to job title

+++ symfony/plugins/orangehrmAdminPlugin/modules/admin/actions/deleteJobTitleAction.class.php
@@ -30,13 +30,15 @@ class deleteJobTitleAction extends sfAction {
     }
 
     public function execute($request) {
+        $jobTitlePermissions = $this->getDataGroupPermissions('job_titles');
+        if ($jobTitlePermissions->canDelete()) {
+            $toBeDeletedJobTitleIds = $request->getParameter('chkSelectRow');
 
-        $toBeDeletedJobTitleIds = $request->getParameter('chkSelectRow');
-
-        if (!empty($toBeDeletedJobTitleIds)) {
-            $this->getJobTitleService()->deleteJobTitle($toBeDeletedJobTitleIds);
-            $this->getUser()->setFlash('success', __(TopLevelMessages::DELETE_SUCCESS));
-            $this->redirect('admin/viewJobTitleList');
+            if (!empty($toBeDeletedJobTitleIds)) {
+                $this->getJobTitleService()->deleteJobTitle($toBeDeletedJobTitleIds);
+                $this->getUser()->setFlash('success', __(TopLevelMessages::DELETE_SUCCESS));
+                $this->redirect('admin/viewJobTitleList');
+            }
         }
     }
